refactor task into another file and make chatbot an object
